refactor(scan): extract downscaleImage into app/lib/image.client
