+++ app/routes/_authed.vehicles.$vehicleId.scan.tsx
@@ -18,6 +18,7 @@ import {
   label,
   textarea,
 } from "~/components/ui";
+import { downscaleImage } from "~/lib/image.client";
 import { requireVehicleAccess } from "~/models/member.server";
 import { extractReceiptScan } from "~/scan/extract.server";
 import { createLogWithScan } from "~/scan/import.server";
@@ -129,37 +130,6 @@ const saveScanFn = createServerFn({ method: "POST" })
     return { vehicleId, logId: log.id };
   });
 
-/**
- * Shrink a phone photo before upload: receipts read fine at 1600px and the
- * extraction round-trip drops from ~10 MB to a few hundred KB. Falls back to
- * the original file when the browser can't decode it (e.g. HEIC outside
- * Safari) — the server cap still applies.
- */
-async function downscaleForUpload(file: File): Promise<File> {
-  const MAX_DIM = 1600;
-  try {
-    const bitmap = await createImageBitmap(file);
-    const scale = Math.min(1, MAX_DIM / Math.max(bitmap.width, bitmap.height));
-    if (scale === 1 && file.type === "image/jpeg") return file;
-
-    const canvas = document.createElement("canvas");
-    canvas.width = Math.max(1, Math.round(bitmap.width * scale));
-    canvas.height = Math.max(1, Math.round(bitmap.height * scale));
-    const ctx = canvas.getContext("2d");
-    if (!ctx) return file;
-    ctx.drawImage(bitmap, 0, 0, canvas.width, canvas.height);
-
-    const blob = await new Promise<Blob | null>((res) =>
-      canvas.toBlob(res, "image/jpeg", 0.85),
-    );
-    if (!blob) return file;
-    const base = file.name.replace(/\.\w+$/, "") || "scan";
-    return new File([blob], `${base}.jpg`, { type: "image/jpeg" });
-  } catch {
-    return file;
-  }
-}
-
 export const Route = createFileRoute("/_authed/vehicles/$vehicleId/scan")({
   component: ScanReceipt,
 });
@@ -214,7 +184,10 @@ function ScanReceipt() {
   async function onPhotoPicked(picked: File | undefined) {
     if (!picked) return;
     setExtractError(null);
-    const upload = await downscaleForUpload(picked);
+    const upload = await downscaleImage(picked, {
+      maxDim: 1600,
+      quality: 0.85,
+    });
     setFile(upload);
     setPreview((old) => {
       if (old) URL.revokeObjectURL(old);
